backend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==4.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.2.0
promise==2.3
PyJWT==1.7.1
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
